dsad

+++ js/app.js
@@ -3,7 +3,7 @@ function add(){
     const forma=document.getElementById("forma");
     let operandoA=forma["operandoA"];
     let operandoB=forma["operandoB"];
-    let resultado=parseInt(operandoA.value)+ parseInt(operandoB.value); 
+    let resultado=parseFloat(operandoA.value)+ parseFloat(operandoB.value); 
     if(isNaN(resultado))
         resultado="Operandos vacios";
     document.getElementById("resultado").innerHTML="resultado: "+resultado;
@@ -14,7 +14,7 @@ function substract(){
     const forma=document.getElementById("forma");
     let operandoA=forma["operandoA"];
     let operandoB=forma["operandoB"];
-    let resultado=parseInt(operandoA.value)- parseInt(operandoB.value); 
+    let resultado=parseFloat(operandoA.value)- parseFloat(operandoB.value); 
     if(isNaN(resultado))
         resultado="Operandos vacios";
     document.getElementById("resultado").innerHTML="resultado: "+resultado;
@@ -25,7 +25,7 @@ function multiply(){
     const forma=document.getElementById("forma");
     let operandoA=forma["operandoA"];
     let operandoB=forma["operandoB"];
-    let resultado=parseInt(operandoA.value)* parseInt(operandoB.value); 
+    let resultado=parseFloat(operandoA.value)* parseFloat(operandoB.value); 
     if(isNaN(resultado))
         resultado="Operandos vacios";
     document.getElementById("resultado").innerHTML="resultado: "+resultado;
@@ -36,7 +36,7 @@ function divide(){
     const forma=document.getElementById("forma");
     let operandoA=forma["operandoA"];
     let operandoB=forma["operandoB"];
-    let resultado=parseInt(operandoA.value)/ parseInt(operandoB.value); 
+    let resultado=parseFloat(operandoA.value)/ parseFloat(operandoB.value); 
     if(isNaN(resultado))
         resultado="Operandos vacios";
     document.getElementById("resultado").innerHTML="resultado: "+resultado;
